updated the delete for categories

+++ src/categories/category.service.ts
@@ -7,35 +7,67 @@ import { Repository, InsertResult } from 'typeorm';
 export class CategoryService {
   constructor(
     @InjectRepository(CategoryEntity)
-    readonly categoryEntity: Repository<CategoryEntity>
+    readonly categoryEntityRepo: Repository<CategoryEntity>
   ) { }
 
 
   async getCategories(): Promise<CategoryEntity[]> {
     // throw new Error('No database connection');
 
-    const results = await this.categoryEntity.find();
+    const results = await this.categoryEntityRepo.find();
     return results;
   }
 
 
   async addCategory(payload: any): Promise<CategoryEntity> {
 
-    const response: InsertResult = await this.categoryEntity
+    const response: InsertResult = await this.categoryEntityRepo
       .createQueryBuilder().insert().values(payload).returning('*').execute();
 
     const createdCat: CategoryEntity = response.raw[0];
     return createdCat;
   }
 
-  async deleteCategory(id: any): Promise<any> {
-    const where = 'id = :id';
-    const query = {
-      id,
-    };
-    const response = await this.categoryEntity
-      .createQueryBuilder().delete().where(where, query).execute()
 
-    return response.raw
+
+  async deleteCategory(id: string): Promise<any> {
+
+    try {
+      const category: CategoryEntity[] = await this.categoryEntityRepo
+        .createQueryBuilder().select('*').where('"id" = :id', { id }).execute();
+
+      if (category[0] == null) {
+        return {
+          statusCode: 404,
+          data: {
+            message: `No category found with ID ${id}`
+          }
+        }
+      }
+
+      const where = 'id = :id';
+      const query = {
+        id,
+      };
+      const response = await this.categoryEntityRepo
+        .createQueryBuilder().delete().where(where, query).execute()
+
+
+      return {
+        statusCode: 200,
+        data: {
+          message: `Category ID ${id} successfully deleted`,
+          other: response.raw[0]
+        }
+      }
+    }
+    catch (error) {
+      return {
+        statusCode: 500,
+        data: {
+          message: error.message
+        }
+      }
+    }
   }
 }
